Copy null field values instead of setting value to empty string
field.Value returns an empty string instead of null, which can cause issues with fields like "__Final Renderings".

Using field.Value property instead of field.SetValue() ensures that the field is reset if set to null. Also, it only sets the value if it is different than the existing value.